requirements.txt
streamlit
pandas
numpy
python-dotenv
langchain
langchain-openai
langchain-experimental
tabulate